Added BaseFixture for Fixtures
* Usage of the same container accessor engine that in functionl tests
* Documented
