+++ Tests/BaseFunctionalTest.php
@@ -15,14 +15,12 @@ declare(strict_types=1);
 
 namespace Mmoreram\BaseBundle\Tests;
 
-use Doctrine\Common\Persistence\ObjectManager;
-use Doctrine\Common\Persistence\ObjectRepository;
 use Exception;
+use Mmoreram\BaseBundle\DependencyInjection\BaseContainerAccessor;
 use PHPUnit_Framework_TestCase;
 use RuntimeException;
 use Symfony\Bundle\FrameworkBundle\Console\Application;
 use Symfony\Component\Console\Input\ArrayInput;
-use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\HttpKernel\Client;
 use Symfony\Component\HttpKernel\KernelInterface;
 
@@ -31,6 +29,8 @@ use Symfony\Component\HttpKernel\KernelInterface;
  */
 abstract class BaseFunctionalTest extends PHPUnit_Framework_TestCase
 {
+    use BaseContainerAccessor;
+
     /**
      * @var Application
      *
@@ -44,13 +44,6 @@ abstract class BaseFunctionalTest extends PHPUnit_Framework_TestCase
      * kernel being used
      */
     protected static $kernel;
-
-    /**
-     * @var ContainerInterface
-     *
-     * Container
-     */
-    protected static $container;
 
     /**
      * @var bool
@@ -238,173 +231,6 @@ abstract class BaseFunctionalTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-     * Get container service.
-     *
-     * @param string $serviceName
-     *
-     * @return mixed
-     */
-    public function get(string $serviceName)
-    {
-        return self::$container->get($serviceName);
-    }
-
-    /**
-     * Container has service.
-     *
-     * @param string $serviceName
-     *
-     * @return bool
-     */
-    public function has(string $serviceName) : bool
-    {
-        return self::$container->has($serviceName);
-    }
-
-    /**
-     * Get container parameter.
-     *
-     * @param string $parameterName
-     *
-     * @return mixed
-     */
-    public function getParameter(string $parameterName)
-    {
-        return self::$container->getParameter($parameterName);
-    }
-
-    /**
-     * Get object repository given an entity namespace.
-     *
-     * @param string $entityNamespace
-     *
-     * @return ObjectRepository|null
-     */
-    protected function getObjectRepository(string $entityNamespace) : ? ObjectRepository
-    {
-        return $this
-            ->get('base.object_repository_provider')
-            ->getObjectRepositoryByEntityNamespace(
-                $this->locateEntity($entityNamespace)
-            );
-    }
-
-    /**
-     * Get object manager given an entity namespace.
-     *
-     * @param string $entityNamespace
-     *
-     * @return ObjectManager|null
-     */
-    protected function getObjectManager(string $entityNamespace) : ? ObjectManager
-    {
-        return $this
-            ->get('base.object_manager_provider')
-            ->getObjectManagerByEntityNamespace(
-                $this->locateEntity($entityNamespace)
-            );
-    }
-
-    /**
-     * Get the entity instance with id $id.
-     *
-     * @param string $entityNamespace
-     * @param mixed  $id
-     *
-     * @return object
-     */
-    public function find(
-        string $entityNamespace,
-        $id
-    ) {
-        return $this
-            ->getObjectRepository($this->locateEntity($entityNamespace))
-            ->find($id);
-    }
-
-    /**
-     * Get the entity instance with criteria.
-     *
-     * @param string $entityNamespace
-     * @param array  $criteria
-     *
-     * @return object
-     */
-    public function findOneBy(
-        string $entityNamespace,
-        array $criteria
-    ) {
-        return $this
-            ->getObjectRepository($this->locateEntity($entityNamespace))
-            ->findOneBy($criteria);
-    }
-
-    /**
-     * Get all entity instances.
-     *
-     * @param string $entityNamespace
-     *
-     * @return array
-     */
-    public function findAll($entityNamespace) : array
-    {
-        return $this
-            ->getObjectRepository($this->locateEntity($entityNamespace))
-            ->findAll();
-    }
-
-    /**
-     * Get all entity instances.
-     *
-     * @param string $entityNamespace
-     * @param array  $criteria
-     *
-     * @return array
-     */
-    public function findBy(
-        string $entityNamespace,
-        array $criteria
-    ) : array {
-        return $this
-            ->getObjectRepository($this->locateEntity($entityNamespace))
-            ->findBy($criteria);
-    }
-
-    /**
-     * Clear the object manager tracking of an entity.
-     *
-     * @param string $entityNamespace
-     */
-    public function clear(string $entityNamespace)
-    {
-        $entityNamespace = $this->locateEntity($entityNamespace);
-        $this
-            ->getObjectManager($entityNamespace)
-            ->clear($entityNamespace);
-    }
-
-    /**
-     * Save entity or array of entities.
-     *
-     * @param mixed $entities
-     */
-    protected function save($entities)
-    {
-        if (!is_array($entities)) {
-            $entities = [$entities];
-        }
-
-        foreach ($entities as $entity) {
-            $entityClass = get_class($entity);
-            $entityManager = $this
-                ->get('base.object_manager_provider')
-                ->getObjectManagerByEntityNamespace($entityClass);
-            $entityManager->persist($entity);
-            $entityManager->flush($entity);
-        }
-    }
-
-    /**
      * Get kernel.
      *
      * @return KernelInterface
@@ -428,37 +254,5 @@ abstract class BaseFunctionalTest extends PHPUnit_Framework_TestCase
         }
 
         return $configuration;
-    }
-
-    /**
-     * Get entity locator given a string.
-     *
-     * Available formats:
-     *
-     * MyBundle\Entity\Namespace\User - Namespace
-     * MyBundle:User - Doctrine short alias
-     * my_prefix:user - When using short DoctrineExtraMapping, prefix:name
-     * my_prefix.entity.user.class - When using DoctrineExtraMapping class param
-     *
-     * @param string $entityAlias
-     *
-     * @return string
-     */
-    private function locateEntity($entityAlias)
-    {
-        if (1 === preg_match('/^.*?\\.entity\\..*?\\.class$/', $entityAlias)) {
-            if (self::$container->hasParameter($entityAlias)) {
-                return $this->getParameter($entityAlias);
-            }
-        }
-
-        if (1 === preg_match('/^[^:]+:[^:]+$/', $entityAlias)) {
-            $possibleEntityAliasShortMapping = str_replace(':', '.entity.', $entityAlias . '.class');
-            if (self::$container->hasParameter($possibleEntityAliasShortMapping)) {
-                return $this->getParameter($possibleEntityAliasShortMapping);
-            }
-        }
-
-        return $entityAlias;
     }
 }
